split up UI components, add AddCommentDialog
